+++ dum/Dialog.hxx
@@ -140,7 +140,7 @@ class Dialog
       CallID mCallId;
       
       // used to capture the 2xx expiration value for the initial subscription response
-      UInt64 mDefaultSubExpiration;
+      UInt32 mDefaultSubExpiration;
 
       // store until we get a response (non-401/407)
       // !jf! this shouldn't be necessary
